Combine Dialog and Say Dialog classes
