--- Assets/Fungus/Narrative/Scripts/Dialog.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b7e8f397d6557484f91f9992f702cff5
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
