--- Assets/Fungus/Narrative/Scripts/DialogText.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4aada7218611f4257bddea1cd4ab8fcf
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
